+++ R/llm.R
@@ -326,6 +326,9 @@ llm_model_list <- function(platform = NULL) {
 
   # get models and ignore errors, add platform name
   models <- lapply(platform, \(p) {
+    if (p != "ollama" && !online()) {
+      return(NULL)
+    }
     tryCatch({
         # skip if google api key isn't set, otherwise it requests login
       if (p %in% c("google_gemini", "google_vertex") &&
@@ -335,7 +338,7 @@ llm_model_list <- function(platform = NULL) {
 
       model_func <- funcs[[p]]
       m <- model_func()
-      cols <- c("platform", names(m))
+      #cols <- c("platform", names(m))
       m$platform <- p
 
       m